🛠️ Refactor: Centralize routing and error handling, enforce domain structure
- Extract history interception logic into `src/lib/router/history.ts`.
- Create centralized error reporter `src/lib/errors/reporter.ts` and route `try/catch` and Svelte `{:catch}` logs through it.
- Move components and models to `src/features/decision-tree/`.
- Move shared elements to `src/lib/ui`, `src/lib/i18n`.
- Create `mise.toml` for tool pinning and CI scripts.
- Document conventions in `AGENTS.md`.
