+++ src/lib/i18n/types.ts
@@ -1,11 +1,4 @@
-export interface DecisionTree {
-    title: i18Led
-    description?: i18Led
-    alternatives?: Record<string, DecisionTree>
-}
-
-export type i18Led = Record<string, string> | string
-
+export type i18Led = Record<string, string> | string;
 
 export function i18nGet(txt: i18Led): string {
     if (typeof txt === 'string') {
